Extract the existing logic for creating a notice template into a function
I didn't write this originally, but I used this
same logic for the file type notice.

So instead of simply copy and pasting it,
this extracts it into a separate function.
